dividing parallel execution further
